average rating

+++ db/migrate/20170203184207_create_beers.rb
@@ -4,7 +4,7 @@ class CreateBeers < ActiveRecord::Migration[5.0]
       t.string :name
       t.string :description
       t.integer :rating
-      t.integer :abv
+      t.float :abv
       t.integer :user_id
       t.integer :style_id
       t.timestamps
